first commit: 校园跑腿小程序+后端

+++ src/main/java/org/example/campusrunnerapi/entity/User.java
@@ -1,19 +1,32 @@
 package org.example.campusrunnerapi.entity;
 
-/**
- * 示例实体类：User
- * 对应数据库表：user
- */
 public class User {
-    private Long id;
+    private Integer id;
     private String username;
     private String password;
 
     // Getter & Setter
-    public Long getId() { return id; }
-    public void setId(Long id) { this.id = id; }
-    public String getUsername() { return username; }
-    public void setUsername(String username) { this.username = username; }
-    public String getPassword() { return password; }
-    public void setPassword(String password) { this.password = password; }
+    public Integer getId() {
+        return id;
+    }
+
+    public void setId(Integer id) {
+        this.id = id;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
 }
